refactor: Finished moving the search.rs file to a separate module.
